WARNING: This commit is rather huge in terms of how many bugs it might create
This rev adds proper support for saving of decaying items by making a optional 6th argument to save and restore that upon __getstate__.
We now delete decay data when items have been decayed, resulting in less memory usage, note that you can't delete a itemattribute loaded from db, such as 'del item.stackable' (unless you first change the stackable state).
We now store the decayPosition directly on the item, this allows us to move the item around using our regular functions and still keep decay support, this fix a bug that caused items moved on to the ground and back into inventory never to decay.
We now store the decayCreature directly into the item, allowing us to decay items that change owner. This cause a little extra work for some internal functions but who cares, it caused for perhaps 1/100 of a milisec. This value is ofcource reset on unpickle of inventory and removed when the item is properly decayed.
Positions etc gets changed by engine calls, and by handleMoveItem in protocol.

This revs need testing to decay behavior.